Some more modularization
